requirements.txt
Pillow==6.0.0
python-barcode==0.10.0
#python-barcode[images]==0.10.0
